add for never died service

+++ WifiSocket/app/src/main/java/com/android/wifi/socket/service/WifiConnectService.java
@@ -1,5 +1,8 @@
 package com.android.wifi.socket.service;
 
+import android.app.AlarmManager;
+import android.app.Notification;
+import android.app.PendingIntent;
 import android.app.Service;
 import android.content.BroadcastReceiver;
 import android.content.Context;
@@ -7,11 +10,15 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
+import android.os.Binder;
 import android.os.Bundle;
 import android.os.IBinder;
+import android.os.SystemClock;
 import android.util.Log;
-import android.widget.Toast;
 
+import com.android.wifi.socket.Activity.MainActivity;
+import com.android.wifi.socket.recieve.WifiBroadcastReciever;
+import com.android.wifi.socket.wifisocket.R;
 import com.android.wifi.socket.wifisocket.WifiAdmin;
 
 /**
@@ -22,6 +29,7 @@ public class WifiConnectService extends Service {
     private WifiAdmin wifiAdmin;
     private ConnectivityManager connectivityManager;
     private int CurrentSSID = 0;
+    private Boolean remove_;
     private BroadcastReceiver mReceiver = new BroadcastReceiver() {
         @Override
         public void onReceive(Context context, Intent intent) {
@@ -32,13 +40,11 @@ public class WifiConnectService extends Service {
                 NetworkInfo.State state = connectivityManager.getNetworkInfo(ConnectivityManager.TYPE_WIFI).getState();
                 Log.e(TAG,"state.toString() = "+ state.toString());
                 if(state.toString().equals("CONNECTED")){
-                    Toast.makeText(context,"CONNECTED",Toast.LENGTH_LONG).show();
                     Log.e(TAG, "CONNECTED");
                 }else if(state.toString().equals("DISCONNECTED")){
                     Log.e(TAG, "Delete CurrentSSID");
-                    Toast.makeText(context,"DISCONNECTED",Toast.LENGTH_LONG).show();
                     wifiAdmin = new WifiAdmin(context);
-                    if(CurrentSSID!=0) {
+                    if(CurrentSSID!=0 && remove_) {
                         deleteWifiConfig(CurrentSSID);
                     }
                 }else{
@@ -53,13 +59,21 @@ public class WifiConnectService extends Service {
     }
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
-        flags = START_REDELIVER_INTENT;
+        Intent notificationIntent = new Intent(this, MainActivity.class);
+        PendingIntent pendingIntent = PendingIntent.getActivity(this, 0, notificationIntent, 0);
+        Notification noti = new Notification.Builder(this)
+                .setContentTitle("Title")
+                .setContentText("Message")
+                .setSmallIcon(R.mipmap.ic_launcher)
+                .setContentIntent(pendingIntent)
+                .build();
+        startForeground(12346, noti);
+
+        flags = Service.START_STICKY;
         Bundle bundle = intent.getExtras();
         CurrentSSID = bundle.getInt("IsInWIfiConfig");
         Log.e(TAG, "CurrentSSID=" + CurrentSSID);
-        if(CurrentSSID==0){
-            this.stopSelf();
-        }
+
         return super.onStartCommand(intent, flags, startId);
     }
     @Override
@@ -68,18 +82,55 @@ public class WifiConnectService extends Service {
         IntentFilter mFilter = new IntentFilter();
         mFilter.addAction(ConnectivityManager.CONNECTIVITY_ACTION);
         registerReceiver(mReceiver, mFilter);
+
+        Intent intent = new Intent(this, WifiBroadcastReciever.class);
+        Bundle bundle = new Bundle();
+        bundle.putInt("IsInWIfiConfig",CurrentSSID);
+        intent.putExtras(bundle);
+        intent.setAction("arui.alarm.action");
+        PendingIntent sender = PendingIntent.getBroadcast(this, 0,
+                intent, 0);
+        long firstime = SystemClock.elapsedRealtime();
+        AlarmManager am = (AlarmManager) this
+                .getSystemService(Context.ALARM_SERVICE);
+
+        // 10秒一个周期，不停的发送广播
+        am.setRepeating(AlarmManager.ELAPSED_REALTIME_WAKEUP, firstime,
+                10 * 1000, sender);
     }
 
     @Override
     public void onDestroy() {
         super.onDestroy();
-        unregisterReceiver(mReceiver);
+        stopForeground(true);
+
+        Bundle bundle = new Bundle();
+        bundle.putInt("IsInWIfiConfig",CurrentSSID);
+        Intent intent=new Intent(this,WifiConnectService.class);
+        intent.putExtras(bundle);
+        startService(intent);
     }
     public void deleteWifiConfig(int networkId){
         boolean flag = wifiAdmin.getWifiManager().removeNetwork(networkId);
         Log.e(TAG,"flag= " +flag+";networkid = "+ networkId);
         Log.e(TAG,"CurrentSSID="+CurrentSSID);
-//        this.stopService(new Intent(this,WifiConnectService.class));
-        this.stopSelf();
+        if(flag){
+            remove_ =false;
+        }
+    }
+    public class MyBinder extends Binder {
+
+        public WifiConnectService getService(){
+            return WifiConnectService.this;
+        }
+    }
+
+    private MyBinder myBinder = new MyBinder();
+
+    public void MyMethod(){
+        for(int i = 0; i < 100; i++)
+        {
+            Log.i(TAG, "BindService-->MyMethod()");
+        }
     }
 }
